requirements.txt
Flask==3.0.0
black==23.11.0
flask_wtf==1.2.1
wtforms_components==0.10.5
simplejson==3.19.2
python-dotenv==1.0.0
flask_sqlalchemy==3.1.1
pytest==7.4.3
gunicorn==21.2.0
requests==2.31.0
google-cloud-datastore==2.21.0
